Added test file, modified template for title
The test documents are actual documents pulled from Solr. It's simply a
copy paste from the query panel, and then stored as an array. I'll get
rid of them eventually, or store them in a better location.

To select the h3 tag, also had to do some modification with the
selectors.