Changed URL in request
Now that the proxy is up and running, we connect to port 8080.  It also lists
out the full path of the resource, instead of just the address.  Even though
they would resolve to the same thing, it's better to be clear of what it is
that they should be accessing.