Added button handler for searching; abstracted search function
Added a handler for the query box so that when the enter key is pressed a search is performed.

Also added shortcut function called simpleSearch, that performs a search function, but with all
of the relevant parameters filled out, so it can be more easily called inline.  This is used in both
the click and key handlers.